+++ internal/css_printer/css_printer_test.go
@@ -301,26 +301,6 @@ func TestAtCharset(t *testing.T) {
 	expectPrintedMinify(t, "@charset \"UTF-8\";", "@charset \"UTF-8\";")
 }
 
-func TestAtNamespace(t *testing.T) {
-	expectPrinted(t, "@namespace\"http://www.com\";", "@namespace \"http://www.com\";\n")
-	expectPrinted(t, "@namespace \"http://www.com\";", "@namespace \"http://www.com\";\n")
-	expectPrinted(t, "@namespace url(http://www.com);", "@namespace \"http://www.com\";\n")
-	expectPrinted(t, "@namespace url(\"http://www.com\");", "@namespace \"http://www.com\";\n")
-	expectPrinted(t, "@namespace ns\"http://www.com\";", "@namespace ns \"http://www.com\";\n")
-	expectPrinted(t, "@namespace ns \"http://www.com\";", "@namespace ns \"http://www.com\";\n")
-	expectPrinted(t, "@namespace ns url(http://www.com);", "@namespace ns \"http://www.com\";\n")
-	expectPrinted(t, "@namespace ns url(\"http://www.com\");", "@namespace ns \"http://www.com\";\n")
-
-	expectPrintedMinify(t, "@namespace\"http://www.com\";", "@namespace\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace \"http://www.com\";", "@namespace\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace url(http://www.com);", "@namespace\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace url(\"http://www.com\");", "@namespace\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace ns\"http://www.com\";", "@namespace ns\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace ns \"http://www.com\";", "@namespace ns\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace ns url(http://www.com);", "@namespace ns\"http://www.com\";")
-	expectPrintedMinify(t, "@namespace ns url(\"http://www.com\");", "@namespace ns\"http://www.com\";")
-}
-
 func TestAtImport(t *testing.T) {
 	expectPrinted(t, "@import\"foo.css\";", "@import \"foo.css\";\n")
 	expectPrinted(t, "@import \"foo.css\";", "@import \"foo.css\";\n")